fix UTs
